[UPD] Moved file with enum of file exts
